before/aftereach in global test file, instead of every test script
