+++ playwright-tests/tests/system-updates/update-infinex-sputnikdao.spec.js
@@ -7,7 +7,6 @@ import {
   SPUTNIK_DAO_FACTORY_ID,
 } from "../../util/sandboxrpc";
 import crypto from "crypto";
-import { cacheCDN } from "../../util/test";
 
 test("update infinex.sputnik-dao.near", async ({ page }) => {
   test.setTimeout(120_000);
@@ -15,8 +14,6 @@ test("update infinex.sputnik-dao.near", async ({ page }) => {
   const daoContractId = `${daoName}.${SPUTNIK_DAO_FACTORY_ID}`;
   const web4ContractId = "treasury-infinex.near";
   const socialNearContractId = "social.near";
-
-  await cacheCDN(page);
 
   const worker = await Worker.init();
