+++ playwright-tests/tests/system-updates/update-policy.spec.js
@@ -1,5 +1,5 @@
 import { expect } from "@playwright/test";
-import { cacheCDN, test } from "../../util/test.js";
+import { test } from "../../util/test.js";
 import {
   DEFAULT_WIDGET_REFERENCE_ACCOUNT_ID,
   SandboxRPC,
@@ -12,8 +12,6 @@ test("should update sputnik-dao policy and upgrade instance with it", async ({
   page,
 }) => {
   test.setTimeout(120_000);
-
-  await cacheCDN(page);
 
   const sandbox = new SandboxRPC();
   await sandbox.init();
