+++ playwright-tests/tests/system-updates/update-specfied-instances.spec.js
@@ -3,19 +3,6 @@ import { test } from "../../util/test.js";
 import { redirectWeb4 } from "../../util/web4.js";
 import { setPageAuthSettings } from "../../util/sandboxrpc.js";
 import { KeyPairEd25519 } from "near-workspaces";
-import fs from "fs";
-import path from "path";
-
-test.afterEach(async ({ page }, testInfo) => {
-  await page.unrouteAll({ behavior: "ignoreErrors" });
-  const video = await page.video();
-  if (video) {
-    const titleFile = testInfo.outputPath("test-title.txt");
-    
-    await fs.promises.writeFile(titleFile, testInfo.title);
-    console.log(`Written video title file ${titleFile}`);
-  }
-});
 
 test("should show system-update for targeted instance", async ({ page }) => {
   const contractId = "treasury-testing.near";
@@ -118,5 +105,5 @@ test("should show system-update if no target instances specified", async ({
   await page.getByRole("link", { name: "Review" }).click();
   await expect(
     page.getByText("Update to latest sputnik-dao contract")
-  ).toBeVisible();
+  ).toBeVisible({ timeout: 15_000 });
 });
