+++ playwright-tests/tests/system-updates/upgrade-instance.spec.js
@@ -1,5 +1,5 @@
 import { expect } from "@playwright/test";
-import { cacheCDN, test } from "../../util/test.js";
+import { test } from "../../util/test.js";
 import {
   DEFAULT_WIDGET_REFERENCE_ACCOUNT_ID,
   SandboxRPC,
@@ -12,8 +12,6 @@ test("should update treasury factory with new web4 contract and self upgrade ins
   page,
 }) => {
   test.setTimeout(150_000);
-
-  await cacheCDN(page);
 
   const sandbox = new SandboxRPC();
   await sandbox.init();
